4.10.4 + Oculus SDK 1.2.1 + Oculus Mobile 1.0.1

+++ Engine/Source/Runtime/Core/Public/Misc/AutomationTest.h
@@ -330,7 +330,7 @@ protected:
 	}
 
 	/** For timers, track the first time this ticks */
-	float StartTime;
+	double StartTime;
 
 	friend class FAutomationTestFramework;
 };
